mysite/requirements.txt
asgiref==3.3.1
Django==3.1.5
django-taggit==1.3.0
pytz==2020.5
sqlparse==0.4.1
psycopg2-binary==2.8.6
django-simple-history==2.12.0